requirements.txt
accelerate==1.1.1
beautifulsoup4==4.12.3
chromadb==0.5.18
langchain==0.3.7
langchain-chroma==0.1.4
langchain-community==0.3.5
langchain-ollama==0.2.0
nltk==3.9.1
ollama==0.3.3
pillow==11.0.0
pypdf==5.1.0
streamlit==1.40.0
tiktoken==0.8.0
